* Move files to reduce dependencies * Remove unused files
