+++ Geometry/HcalCommonData/BuildFile.xml
@@ -1,5 +1,6 @@
 <use   name="DetectorDescription/Base"/>
 <use   name="DetectorDescription/Core"/>
+<use   name="Geometry/CaloTopology"/>
 <export>
   <lib   name="1"/>
 </export>
