+++ Geometry/HcalCommonData/plugins/BuildFile.xml
@@ -1,8 +1,9 @@
-<library   file="*.cc" name="GeometryHcalTowerAlgoPlugins">
+<library   file="*.cc" name="GeometryHcalCommonDataPlugins">
   <use   name="FWCore/Framework"/>
   <use   name="FWCore/ParameterSet"/>
   <use   name="Geometry/Records"/>
-  <use   name="Geometry/HcalTowerAlgo"/>
+  <use   name="Geometry/HcalCommonData"/>
+  <use   name="CondFormats/GeometryObjects"/>
   <use   name="boost"/>
   <flags   EDM_PLUGIN="1"/>
 </library>
