+++ Geometry/HcalTowerAlgo/BuildFile.xml
@@ -3,7 +3,6 @@
 </export>
 <use   name="Geometry/CaloGeometry"/>
 <use   name="Geometry/CaloTopology"/>
-<use   name="Geometry/HcalCommonData"/>
 <use   name="DataFormats/HcalDetId"/>
 <use   name="DataFormats/CaloTowers"/>
 <use   name="FWCore/MessageLogger"/>
