+++ Geometry/HcalTowerAlgo/test/BuildFile.xml
@@ -15,7 +15,3 @@
 <library   file="HcalGeometryDetIdAnalyzer.cc" name="testHcalGeometryDetIdESProducer">
  <flags   EDM_PLUGIN="1"/>
 </library>
-<library   file="HcalParametersAnalyzer.cc" name="testHcalGeometryESProducer">
- <use name="CondFormats/GeometryObjects"/>
- <flags   EDM_PLUGIN="1"/>
-</library>
