--- Geometry/Records/src/ES_MuonDDDConstants.cc
@@ -1,4 +1,0 @@
-#include "Geometry/MuonNumbering/interface/MuonDDDConstants.h"
-#include "FWCore/Utilities/interface/typelookup.h"
-
-TYPELOOKUP_DATA_REG(MuonDDDConstants);
