+++ Geometry/Records/test/BuildFile.xml
@@ -1,0 +1,9 @@
+<use   name="FWCore/Framework"/>
+<use   name="FWCore/ParameterSet"/>
+<use   name="FWCore/PythonParameterSet"/>
+<use   name="FWCore/Utilities"/>
+<use   name="Geometry/Records"/>
+<library   file="HcalParametersAnalyzer.cc" name="testHcalParametersAnalyzer">
+ <use name="CondFormats/GeometryObjects"/>
+ <flags   EDM_PLUGIN="1"/>
+</library>
